Refactoring override omniauth callbacks controller and move to api folder.
