+++ config/initializers/devise_token_auth.rb
@@ -24,7 +24,7 @@ DeviseTokenAuth.setup do |config|
   # This route will be the prefix for all oauth2 redirect callbacks. For
   # example, using the default '/omniauth', the github oauth2 provider will
   # redirect successful authentications to '/omniauth/github/callback'
-  config.omniauth_prefix = "/api/v1/auth"
+  config.omniauth_prefix = "/api/auth"
 
   # By default sending current password is not needed for the password update.
   # Uncomment to enforce current_password param to be checked before all
